MetaBoxes: Moving meta boxes out of the generic editor module
