+++ edit-post/assets/stylesheets/_z-index.scss
@@ -15,8 +15,8 @@ $z-layers: (
 	'.editor-inserter__tabs': 1,
 	'.editor-inserter__tab.is-active': 1,
 	'.components-panel__header': 1,
-	'.editor-meta-boxes-area.is-loading:before': 1,
-	'.editor-meta-boxes-area .spinner': 2,
+	'.edit-post-meta-boxes-area.is-loading:before': 1,
+	'.edit-post-meta-boxes-area .spinner': 2,
 	'.blocks-format-toolbar__link-modal': 2,
 	'.editor-block-contextual-toolbar': 2,
 	'.editor-block-switcher__menu': 2,
