+++ edit-post/components/layout/style.scss
@@ -62,7 +62,7 @@
 	padding: 10px 0 10px;
 	clear: both;
 
-	.editor-meta-boxes-area {
+	.edit-post-meta-boxes-area {
 		max-width: $visual-editor-max-width;
 		margin: auto;
 	}
