+++ edit-post/index.js
@@ -19,6 +19,7 @@ import './assets/stylesheets/main.scss';
 import './hooks';
 import Layout from './components/layout';
 import store from './store';
+import { initializeMetaBoxState } from './store/actions';
 
 export * from './api';
 
@@ -91,7 +92,7 @@ export function initializeEditor( id, post, settings ) {
 	const reboot = reinitializeEditor.bind( null, target, settings );
 	const ReduxProvider = createProvider( 'edit-post' );
 
-	const provider = render(
+	render(
 		<EditorProvider settings={ settings } post={ post }>
 			<ErrorBoundary onError={ reboot }>
 				<ReduxProvider store={ store }>
@@ -103,6 +104,8 @@ export function initializeEditor( id, post, settings ) {
 	);
 
 	return {
-		initializeMetaBoxes: provider.initializeMetaBoxes,
+		initializeMetaBoxes( metaBoxes ) {
+			store.dispatch( initializeMetaBoxState( metaBoxes ) );
+		},
 	};
 }
