+++ edit-post/store/test/actions.js
@@ -11,6 +11,8 @@ import {
 	togglePublishSidebar,
 	setViewportType,
 	toggleFeature,
+	requestMetaBoxUpdates,
+	initializeMetaBoxState,
 } from '../actions';
 
 describe( 'actions', () => {
@@ -94,6 +96,29 @@ describe( 'actions', () => {
 			expect( toggleFeature( feature ) ).toEqual( {
 				type: 'TOGGLE_FEATURE',
 				feature,
+			} );
+		} );
+	} );
+
+	describe( 'requestMetaBoxUpdates', () => {
+		it( 'should return the REQUEST_META_BOX_UPDATES action', () => {
+			expect( requestMetaBoxUpdates() ).toEqual( {
+				type: 'REQUEST_META_BOX_UPDATES',
+			} );
+		} );
+	} );
+
+	describe( 'initializeMetaBoxState', () => {
+		it( 'should return the META_BOX_STATE_CHANGED action with a hasChanged flag', () => {
+			const metaBoxes = {
+				side: true,
+				normal: true,
+				advanced: false,
+			};
+
+			expect( initializeMetaBoxState( metaBoxes ) ).toEqual( {
+				type: 'INITIALIZE_META_BOX_STATE',
+				metaBoxes,
 			} );
 		} );
 	} );
